# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ÖzelGüvenlik.Online API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
  /** @nullable */
  message?: string | null;
}

export interface User {
  id: number;
  username: string;
  email: string;
  role: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  nameColor?: string | null;
  nameAnimated?: boolean;
  isBanned?: boolean;
  /** @nullable */
  banReason?: string | null;
  /** @nullable */
  banExpiresAt?: string | null;
  createdAt: string;
}

export interface UserProfile {
  id: number;
  username: string;
  role: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  nameColor?: string | null;
  nameAnimated?: boolean;
  listingCount?: number;
  createdAt: string;
}

export interface RegisterInput {
  /** @minLength 3 */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface ProfileUpdate {
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface Listing {
  id: number;
  title: string;
  company: string;
  city: string;
  /** @nullable */
  salary?: string | null;
  workType?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  requirements?: string | null;
  status: string;
  viewCount: number;
  likeCount: number;
  isFeatured?: boolean;
  /** @nullable */
  applyUrl?: string | null;
  /** @nullable */
  companyLogoUrl?: string | null;
  /** @nullable */
  authorId?: number | null;
  /** @nullable */
  authorUsername?: string | null;
  isLikedByMe?: boolean;
  isFavoritedByMe?: boolean;
  createdAt: string;
}

export interface ListingInput {
  title: string;
  company: string;
  city: string;
  /** @nullable */
  salary?: string | null;
  workType?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  requirements?: string | null;
  /** @nullable */
  applyUrl?: string | null;
  /** @nullable */
  companyLogoUrl?: string | null;
}

export interface ListingUpdate {
  title?: string;
  company?: string;
  city?: string;
  /** @nullable */
  salary?: string | null;
  workType?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  requirements?: string | null;
  status?: string;
  /** @nullable */
  applyUrl?: string | null;
  isFeatured?: boolean;
}

export interface ListingsResponse {
  listings: Listing[];
  total: number;
  page: number;
  limit: number;
}

export interface CityCount {
  city: string;
  count: number;
}

export interface ListingStats {
  total: number;
  today: number;
  featured: number;
  byCity: CityCount[];
}

export interface LikeResponse {
  liked: boolean;
  likeCount: number;
}

export interface FavoriteResponse {
  favorited: boolean;
}

export interface ChatMessage {
  id: number;
  content: string;
  userId: number;
  username: string;
  /** @nullable */
  userAvatarUrl?: string | null;
  /** @nullable */
  userNameColor?: string | null;
  userNameAnimated?: boolean;
  userRole?: string;
  /** @nullable */
  replyToId?: number | null;
  /** @nullable */
  replyToUsername?: string | null;
  /** @nullable */
  replyToContent?: string | null;
  isPinned?: boolean;
  mentions?: string[];
  createdAt: string;
}

export interface ChatMessageInput {
  /** @minLength 1 */
  content: string;
  /** @nullable */
  replyToId?: number | null;
}

export interface OnlineCount {
  count: number;
  fakeBonus?: number;
}

export interface Notification {
  id: number;
  type: string;
  message: string;
  isRead: boolean;
  /** @nullable */
  linkUrl?: string | null;
  createdAt: string;
}

export interface UnreadCount {
  count: number;
}

export interface Announcement {
  id: number;
  content: string;
  isActive: boolean;
  createdAt: string;
}

export interface AnnouncementInput {
  content: string;
}

export interface AdminStats {
  totalUsers: number;
  onlineUsers: number;
  totalListings: number;
  todayListings: number;
  totalMessages: number;
  bannedUsers: number;
  pendingListings: number;
  reportedMessages: number;
}

export interface AdminUsersResponse {
  users: User[];
  total: number;
}

export interface BanInput {
  reason: string;
  /** @nullable */
  expiresAt?: string | null;
}

export interface RoleUpdate {
  role: string;
}

export interface NameColorUpdate {
  /** @nullable */
  nameColor?: string | null;
  nameAnimated?: boolean;
}

export interface AdminSettings {
  chatLocked: boolean;
  fakeOnlineBonus: number;
  maintenanceMode: boolean;
  /** @nullable */
  welcomeMessage?: string | null;
}

export interface AdminSettingsUpdate {
  chatLocked?: boolean;
  fakeOnlineBonus?: number;
  maintenanceMode?: boolean;
  /** @nullable */
  welcomeMessage?: string | null;
}

export interface FakeLikesInput {
  count: number;
}

export interface BannedWord {
  id: number;
  word: string;
  createdAt: string;
}

export interface BannedWordInput {
  word: string;
}

export type GetListingsParams = {
page?: number;
limit?: number;
city?: string;
search?: string;
featured?: boolean;
};

export type GetChatMessagesParams = {
limit?: number;
before?: number;
};

export type GetAdminUsersParams = {
page?: number;
search?: string;
};