# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ÖzelGüvenlik.Online API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
...
  /chat/messages:
    get:
      operationId: getChatMessages
      tags: [chat]
      summary: Get recent chat messages
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
        - name: before
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"

    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a chat message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
  /chat/messages/{id}:
    delete:
      operationId: deleteChatMessage
      tags: [chat]
      summary: Delete a chat message (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /chat/messages/{id}/pin:
    post:
      operationId: pinChatMessage
      tags: [chat]
      summary: Pin/unpin a chat message (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pin toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /announcements:
    get:
      operationId: getAnnouncements
      tags: [announcements]
      summary: Get active announcements (kayan yazılar)
      responses:
        "200":
          description: Announcements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"

    post:
      operationId: createAnnouncement
      tags: [announcements]
      summary: Create announcement (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /announcements/{id}:
    delete:
      operationId: deleteAnnouncement
      tags: [announcements]
      summary: Delete announcement (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
    User:
      type: object
      required: [id, username, email, role, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        nameColor:
          type: ["string", "null"]
        nameAnimated:
          type: boolean
        isBanned:
          type: boolean
        banReason:
          type: ["string", "null"]
        banExpiresAt:
          type: ["string", "null"]
        createdAt:
          type: string
...
    RegisterInput:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 3
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
...
    ChatMessage:
      type: object
      required: [id, content, userId, username, createdAt]
      properties:
        id:
          type: integer
        content:
          type: string
        userId:
          type: integer
        username:
          type: string
        userAvatarUrl:
          type: ["string", "null"]
        userNameColor:
          type: ["string", "null"]
        userNameAnimated:
          type: boolean
        userRole:
          type: string
        replyToId:
          type: ["integer", "null"]
        replyToUsername:
          type: ["string", "null"]
        replyToContent:
          type: ["string", "null"]
        isPinned:
          type: boolean
        mentions:
          type: array
          items:
            type: string
        createdAt:
          type: string
    ChatMessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
        replyToId:
          type: ["integer", "null"]
...
    Announcement:
      type: object
      required: [id, content, isActive, createdAt]
      properties:
        id:
          type: integer
        content:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
    AnnouncementInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
...