# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ÖzelGüvenlik.Online API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  username: string;
  email: string;
  role: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  nameColor?: string | null;
  nameAnimated?: boolean;
  isBanned?: boolean;
  /** @nullable */
  banReason?: string | null;
  /** @nullable */
  banExpiresAt?: string | null;
  createdAt: string;
}